+++ client/go/internal/admin/vespa-wrapper/services/stop.go
