+++ client/go/internal/cli/cmd/vespa/main.go
@@ -1,5 +1,5 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-// Cobra commands main file
+// Entrypoint for Vespa CLI
 // Author: bratseth
 
 package main
